Streamlined the data passage on admin page.
There is now only one file for dummy data and it successfully is passed
through all the components on the admin page and work as intended.

I also fixed a bug with the DataTable css causing Tables inside of extend
components to have borders.

Fixed a bug with the ScheduleExpand where the selected map marker
wouldn't update correctly when a map marker is deleted.

Commented out the icon changes in the AdminMap since I've decided I
don't like them.