Admin page now calls server to get the semester.
Also implemented the Activate/Deactivate buttons on the admin subheader.

Fixed a bug with routing where the Node server wouldn't allow url
redirects to go to the admin page.

Fixed the outdated model for Events on the server. The "time" array was
replaced with "startTime" and "endTime" instead. This change also was
made to all existing code that utilized the "time" field, so every
"event" is now consistant through the entire application.