Changed admin header bars to have tabs.
This is to make room for the importer component, which previously had
no place in the UI other than a button. Since it will require more than
that, this change has been made to give its own space.

Note: A nice thing to implement would be some feedback in the "Create
New" tabs for the user to know if the operation was successful or not.
This may be done later if I have time.