Fixed Bootstrap Accordion bug
I found this bug when I updated Bootstrap from 4.0.0 to 4.2.1 while
implementing spinners. I fixed this bug by changing the "id" values of
the broken accordions so they start with the text "id" instead of just
starting with the object id, which had the potential to start with a
digit instead of a character.

See this for details:
https://stackoverflow.com/questions/37270787/uncaught-syntaxerror-
failed-to-execute-queryselector-on-document