Added feedback from admin operations.
Now when creating or updating any data object such as I-Teams, colleges,
or events, feedback from the server will be displayed to the user.

I also noticed an oversight with the UpdateColleges function on the
server. It now properly checks if the college name exists already and
will return an error message if it does already exist.