Fixed Google Maps bugs
Fixed a zoom bug with the mobile-sized Google Maps. It used to zoom
as far in as possible when there was only one MapSpot, making it a
potential bother for the user. It now won't zoom in past the size of
the Manwaring Center.

The other bug was when the desk-top sized map would suddenly go to the
ocean when an event without any mapSpots were selected. This was an
issue of the data containing an empty array for MapSpots instead of
Null. The Map component now properly handles both situations.